Implement basic rest API functionality
I extended predefined viewsets to add listing and addition functionality
to the api app. It is barebones for now but it is a good foundation. I
also decided to use routers from rest_framework instead of default
django mechanisms for urls as they seem to offer more basic
functionalities which I probably will not extend or need but which could
serve as a useful addition if they are already built-in. For example
with url patterns /api/ did not resolve to anything (only /api/notes/
worked) but with the default router it also shows something.