Add filtering notes by their tags
A lot has changed between last commit and this one but many of the
changes are purely cosmetic (like sorting tags by their names,
displaying tags alongside notes or boarders around elements thanks to
addition of basic CSS).

But the most important part of this commit was without a doubt adding a
search bar, it works mostly like I want it to work apart from not
clearing on page refresh and when a new tag is inserted it appears in
the search in alphabetical order which is something I did not plan and
don't know how to change.