Allow adding new tags and start reworking add_note
This commit should be two commits because I forgot to commit yesterday's
changes but they were minor compared to today's so it is not a big deal,
I might rework that later or not. The most important part is that adding
new notes with sources was improved and I believe I am on the right to
finally add them properly but I need to add a custom clean function and
everything will be able to make use of textinput widget in
modelfieldchoice for source field in connection_formset.