Make adding many sources possible
The main point of this commit was to finally allow user to easily add
arbitrarily many sources to note at once. Also includes a small fix to
note.html template because initially it did not display comments (as
there were no comments to begin with) but now that comments for sources
were implemented they can be displayed. For that I had to edit the
views.py file because earlier I only passed connection.source
attributes, not caring about other values in connection itself, so now
pass the connection set so I can get rid of predecessors list.