Fix assigning id to dynamic forms
When clicking delete button on any other place than last note it did not
update indices correctly for all other elements inside form-row after
the deleted one. For example if there were 4 elements and the 3rd was
deleted the 4th one would still have an id of 4 however total number of
forms (django specific attribute) was now at 3 so adding a new source
increased it to 4 and now there were two different forms with the same
id.

The same fix (though shorter) has to be done for tags as well.