Make use of django autocomplete light to add notes
I wanted to have autocomplete functionality in the form for adding new
notes, I managed to get it to work, however dynamically adding new forms
is broken again because this is now much complex than before so I will
have to look in to that soon, to make them work again. The main problem
is now that adding a new field by cloning the previous one increments
the number of select fields by one for no apparent reason (at the time
of writing this commit).