Fix dynamic forms for adding multiple sources
Because I switched to using dal for source field in connection_formsets,
I had to redo adding and deleting them dynamically. Because of dal the
html was much more complex and more things needed to be changed so that
ids were assigned correctly and total number of forms counter was
updated correctly. Removing was difficult a little bit because the ids
needed to be recomputed while also keeping the events tied to the
button.

Only functionality lacking is readding a form when all of them were
deleted however I do not know if I will add that though previously it
was not much work but now it might be different.