Add navigation bar to the page
I added navigation bar to make switching between all functionalities of
the app possible by just clicking on item there, instead of manually
typing url and it works and is styled fine. In addition, I changed font
everywhere to Roboto, not just in the index.html, to add more uniform
and better look.